[packaging] Fix soname used by the libraries. Contributes to JB#12048
The shared libraries has real name as soname and the library symlink
in devel package also points directly to the real library instead of
going through soname symlink.

Change sonames for libraries from "xxx.so.0.2.0" to "xxx.so.0"
and let post-install ldconfig create the soname to realname
symlink.

Make devel package library symlink "xxx.so" point to "xxx.so.0".