[packaging] Make sure the various version numbers stay in sync
There are 6 files that contain package version number and it is too
easy to forget to update some of them.

Add verify_version script that goes through all of the files and
complains if mismatches are found (modified from similar script
in mce tree).

Make rpm-builds fail if version numbers are sufficiently off sync.